+++ src/tomahawk/graphics/d3d11.h
@@ -409,14 +409,14 @@ namespace Tomahawk
 				void FlushState() override;
 				bool Map(ElementBuffer* Resource, ResourceMap Mode, MappedSubresource* Map) override;
 				bool Unmap(ElementBuffer* Resource, MappedSubresource* Map) override;
-				bool UpdateBuffer(ElementBuffer* Resource, void* Data, uint64_t Size) override;
+				bool UpdateBuffer(ElementBuffer* Resource, void* Data, size_t Size) override;
 				bool UpdateBuffer(Shader* Resource, const void* Data) override;
 				bool UpdateBuffer(MeshBuffer* Resource, Compute::Vertex* Data) override;
 				bool UpdateBuffer(SkinMeshBuffer* Resource, Compute::SkinVertex* Data) override;
 				bool UpdateBuffer(InstanceBuffer* Resource) override;
 				bool UpdateBuffer(RenderBufferType Buffer) override;
 				bool UpdateBufferSize(Shader* Resource, size_t Size) override;
-				bool UpdateBufferSize(InstanceBuffer* Resource, uint64_t Size) override;
+				bool UpdateBufferSize(InstanceBuffer* Resource, size_t Size) override;
 				void ClearBuffer(InstanceBuffer* Resource) override;
 				void ClearWritable(Texture2D* Resource) override;
 				void ClearWritable(Texture2D* Resource, float R, float G, float B) override;
@@ -458,7 +458,7 @@ namespace Tomahawk
 				bool GenerateTexture(Texture2D* Resource) override;
 				bool GenerateTexture(Texture3D* Resource) override;
 				bool GenerateTexture(TextureCube* Resource) override;
-				bool GetQueryData(Query* Resource, uint64_t* Result, bool Flush) override;
+				bool GetQueryData(Query* Resource, size_t* Result, bool Flush) override;
 				bool GetQueryData(Query* Resource, bool* Result, bool Flush) override;
 				void QueryBegin(Query* Resource) override;
 				void QueryEnd(Query* Resource) override;
@@ -508,7 +508,7 @@ namespace Tomahawk
 				void* GetDevice() const override;
 				void* GetContext() const override;
 				bool IsValid() const override;
-				bool CreateDirectBuffer(uint64_t Size);
+				bool CreateDirectBuffer(size_t Size);
 				bool CreateTexture2D(Texture2D* Resource, DXGI_FORMAT Format);
 				bool CreateTextureCube(TextureCube* Resource, DXGI_FORMAT Format);
 				ID3D11InputLayout* GenerateInputLayout(D3D11Shader* Shader);
